outsoured add alias dialog
